Function mailchecker::is_valid

pub fn is_valid(email: &str) -> bool

Usage

The mailchecker crate provides a is_valid(email) function that ensure the email is valid and does not come from a disposable email service.

Examples

Emails below will be considered valid:

extern crate mailchecker;

assert_eq!(true, mailchecker::is_valid("plop@plop.com"));
assert_eq!(true, mailchecker::is_valid("my.ok@ok.plop.com"));
assert_eq!(true, mailchecker::is_valid("my+ok@ok.plop.com"));
assert_eq!(true, mailchecker::is_valid("my=ok@ok.plop.com"));
assert_eq!(true, mailchecker::is_valid("ok@gmail.com"));
assert_eq!(true, mailchecker::is_valid("ok@hotmail.com"));

Emails below will be considered invalid because of a wrong format:

extern crate mailchecker;

assert_eq!(false, mailchecker::is_valid(""));
assert_eq!(false, mailchecker::is_valid("plopplop.com"));
assert_eq!(false, mailchecker::is_valid("my+ok@ok=plop.com"));
assert_eq!(false, mailchecker::is_valid("my,ok@ok.plop.com"));
assert_eq!(false, mailchecker::is_valid("  ok@gmail.com  "));
assert_eq!(false, mailchecker::is_valid("  ok@gmail.com"));
assert_eq!(false, mailchecker::is_valid("ok@gmail.com  "));
assert_eq!(false, mailchecker::is_valid("\nok@gmail.com\n"));
assert_eq!(false, mailchecker::is_valid("\nok@gmail.com"));
assert_eq!(false, mailchecker::is_valid("ok@gmail.com\n"));

Emails below will be considered invalid because they came from an invalid domain

extern crate mailchecker;

assert_eq!(false, mailchecker::is_valid("ok@33mail.com"));
assert_eq!(false, mailchecker::is_valid("ok@ok.33mail.com"));
assert_eq!(false, mailchecker::is_valid("ok@guerrillamailblock.com"));